Moving common function from DistGitAPI to GitOperations and Dockerfile
