+++ container_workflow_tool/utility.py
@@ -1,6 +1,7 @@
 import sys
 import argparse
 import os
+import re
 import logging
 
 import textwrap
